videoapp/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.2
google-cloud-texttospeech==2.16.3
python-multipart==0.0.9
ffmpeg-python==0.2.0
Pillow==10.3.0
aiofiles==23.2.1
python-dotenv==1.0.1
httpx==0.27.0
requests==2.31.0
